requirements.txt
toga
psycopg2-binary
bcrypt
python-dotenv
